+++ clients/venice-push-job/src/main/java/com/linkedin/venice/hadoop/VenicePushJob.java
@@ -131,6 +131,7 @@ import com.linkedin.venice.heartbeat.PushJobHeartbeatSender;
 import com.linkedin.venice.heartbeat.PushJobHeartbeatSenderFactory;
 import com.linkedin.venice.jobs.ComputeJob;
 import com.linkedin.venice.jobs.DataWriterComputeJob;
+import com.linkedin.venice.jobs.StageMetricsSnapshot;
 import com.linkedin.venice.message.KafkaKey;
 import com.linkedin.venice.meta.BufferReplayPolicy;
 import com.linkedin.venice.meta.HybridStoreConfig;
@@ -147,7 +148,6 @@ import com.linkedin.venice.schema.writecompute.WriteComputeOperation;
 import com.linkedin.venice.security.SSLFactory;
 import com.linkedin.venice.serialization.avro.AvroProtocolDefinition;
 import com.linkedin.venice.serialization.avro.InternalAvroSpecificSerializer;
-import com.linkedin.venice.spark.datawriter.task.StageMetricsRegistry;
 import com.linkedin.venice.spark.utils.RmdPushUtils;
 import com.linkedin.venice.status.PushJobDetailsStatus;
 import com.linkedin.venice.status.protocol.PushJobDetails;
@@ -1356,7 +1356,7 @@ public class VenicePushJob implements AutoCloseable {
    * job, or empty if the compute engine doesn't support stage metrics or the job has not been initialized.
    */
   @VisibleForTesting
-  public Optional<StageMetricsRegistry.Snapshot> getStageMetricsSnapshot() {
+  public Optional<StageMetricsSnapshot> getStageMetricsSnapshot() {
     return dataWriterComputeJob != null ? dataWriterComputeJob.getStageMetricsSnapshot() : Optional.empty();
   }
