+++ clients/venice-push-job/src/test/java/com/linkedin/venice/spark/datawriter/task/StageMetricsRegistryTest.java
@@ -129,21 +129,6 @@ public class StageMetricsRegistryTest {
         "Report should show dash placeholders for both Records In and Bytes In");
   }
 
-  @Test
-  public void testFormatBytes() {
-    Assert.assertEquals(StageMetricsRegistry.formatBytes(500), "500 B");
-    Assert.assertEquals(StageMetricsRegistry.formatBytes(1024), "1.0 KB");
-    Assert.assertEquals(StageMetricsRegistry.formatBytes(1536), "1.5 KB");
-    Assert.assertEquals(StageMetricsRegistry.formatBytes(1024 * 1024), "1.0 MB");
-    Assert.assertEquals(StageMetricsRegistry.formatBytes(1024L * 1024 * 1024), "1.0 GB");
-    Assert.assertEquals(StageMetricsRegistry.formatBytes(1024L * 1024 * 1024 * 3 / 2), "1.5 GB");
-  }
-
-  @Test
-  public void testFormatNumber() {
-    Assert.assertEquals(StageMetricsRegistry.formatNumber(0), "0");
-    Assert.assertEquals(StageMetricsRegistry.formatNumber(999), "999");
-    Assert.assertEquals(StageMetricsRegistry.formatNumber(1000), "1,000");
-    Assert.assertEquals(StageMetricsRegistry.formatNumber(1234567), "1,234,567");
-  }
+  // Format method tests removed — formatting is validated indirectly via testGenerateReportWithData
+  // which checks formatted values ("1,000", "1.0 MB", "2.5s") in the report output.
 }
