--- internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/endToEnd/TestRepush.java
@@ -1,1599 +1,0 @@
-package com.linkedin.venice.endToEnd;
-
-import static com.linkedin.davinci.store.rocksdb.RocksDBServerConfig.ROCKSDB_BLOCK_CACHE_SIZE_IN_BYTES;
-import static com.linkedin.davinci.store.rocksdb.RocksDBServerConfig.ROCKSDB_PLAIN_TABLE_FORMAT_ENABLED;
-import static com.linkedin.venice.ConfigKeys.CLIENT_SYSTEM_STORE_REPOSITORY_REFRESH_INTERVAL_SECONDS;
-import static com.linkedin.venice.ConfigKeys.CLIENT_USE_SYSTEM_STORE_REPOSITORY;
-import static com.linkedin.venice.ConfigKeys.DATA_BASE_PATH;
-import static com.linkedin.venice.ConfigKeys.DAVINCI_PUSH_STATUS_CHECK_INTERVAL_IN_MS;
-import static com.linkedin.venice.ConfigKeys.PERSISTENCE_TYPE;
-import static com.linkedin.venice.meta.PersistenceType.ROCKS_DB;
-import static com.linkedin.venice.schema.rmd.RmdConstants.TIMESTAMP_FIELD_NAME;
-import static com.linkedin.venice.schema.rmd.v1.CollectionRmdTimestamp.ACTIVE_ELEM_TS_FIELD_NAME;
-import static com.linkedin.venice.schema.rmd.v1.CollectionRmdTimestamp.TOP_LEVEL_TS_FIELD_NAME;
-import static com.linkedin.venice.utils.IntegrationTestPushUtils.createStoreForJob;
-import static com.linkedin.venice.utils.IntegrationTestPushUtils.getSamzaProducer;
-import static com.linkedin.venice.utils.IntegrationTestPushUtils.sendStreamingDeleteRecord;
-import static com.linkedin.venice.utils.IntegrationTestPushUtils.sendStreamingRecord;
-import static com.linkedin.venice.utils.TestUtils.assertCommand;
-import static com.linkedin.venice.utils.TestWriteUtils.STRING_SCHEMA;
-import static com.linkedin.venice.utils.TestWriteUtils.getTempDataDirectory;
-import static com.linkedin.venice.utils.TestWriteUtils.loadFileAsString;
-import static com.linkedin.venice.utils.TestWriteUtils.writeSimpleAvroFileWithCustomSize;
-import static com.linkedin.venice.utils.TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.DATA_WRITER_COMPUTE_JOB_CLASS;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.DEFAULT_KEY_FIELD_PROP;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.DEFAULT_VALUE_FIELD_PROP;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.KAFKA_INPUT_COMBINER_ENABLED;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.KAFKA_INPUT_FABRIC;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.KAFKA_INPUT_MAX_RECORDS_PER_MAPPER;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.KAFKA_INPUT_TOPIC;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.REPUSH_TTL_ENABLE;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.REPUSH_TTL_START_TIMESTAMP;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.REWIND_TIME_IN_SECONDS_OVERRIDE;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.SEND_CONTROL_MESSAGES_DIRECTLY;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.SOURCE_KAFKA;
-import static com.linkedin.venice.vpj.VenicePushJobConstants.VENICE_REPUSH_SOURCE_PUBSUB_BROKER;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertNotEquals;
-import static org.testng.Assert.assertNotNull;
-import static org.testng.Assert.assertNull;
-import static org.testng.Assert.assertTrue;
-
-import com.linkedin.avroutil1.compatibility.AvroCompatibilityHelper;
-import com.linkedin.davinci.client.DaVinciClient;
-import com.linkedin.davinci.client.DaVinciConfig;
-import com.linkedin.davinci.client.StorageClass;
-import com.linkedin.davinci.client.factory.CachingDaVinciClientFactory;
-import com.linkedin.davinci.replication.RmdWithValueSchemaId;
-import com.linkedin.davinci.replication.merge.RmdSerDe;
-import com.linkedin.davinci.replication.merge.StringAnnotatedStoreSchemaCache;
-import com.linkedin.davinci.storage.chunking.SingleGetChunkingAdapter;
-import com.linkedin.davinci.store.StorageEngine;
-import com.linkedin.davinci.store.record.ValueRecord;
-import com.linkedin.venice.client.store.AvroGenericStoreClient;
-import com.linkedin.venice.client.store.ClientConfig;
-import com.linkedin.venice.client.store.ClientFactory;
-import com.linkedin.venice.compression.CompressionStrategy;
-import com.linkedin.venice.controllerapi.ControllerClient;
-import com.linkedin.venice.controllerapi.ControllerResponse;
-import com.linkedin.venice.controllerapi.UpdateStoreQueryParams;
-import com.linkedin.venice.controllerapi.VersionCreationResponse;
-import com.linkedin.venice.exceptions.VeniceException;
-import com.linkedin.venice.hadoop.VenicePushJob;
-import com.linkedin.venice.integration.utils.VeniceClusterWrapper;
-import com.linkedin.venice.integration.utils.VeniceMultiClusterWrapper;
-import com.linkedin.venice.integration.utils.VeniceRouterWrapper;
-import com.linkedin.venice.integration.utils.VeniceServerWrapper;
-import com.linkedin.venice.meta.ReadOnlySchemaRepository;
-import com.linkedin.venice.meta.Store;
-import com.linkedin.venice.meta.Version;
-import com.linkedin.venice.samza.VeniceSystemProducer;
-import com.linkedin.venice.schema.SchemaEntry;
-import com.linkedin.venice.schema.rmd.RmdSchemaEntry;
-import com.linkedin.venice.schema.rmd.RmdSchemaGenerator;
-import com.linkedin.venice.schema.writecompute.DerivedSchemaEntry;
-import com.linkedin.venice.schema.writecompute.WriteComputeSchemaConverter;
-import com.linkedin.venice.spark.datawriter.jobs.DataWriterSparkJob;
-import com.linkedin.venice.spark.datawriter.task.StageMetricsRegistry;
-import com.linkedin.venice.spark.datawriter.task.StageMetricsRegistry.Snapshot.StageSummary;
-import com.linkedin.venice.stats.VeniceMetricsRepository;
-import com.linkedin.venice.utils.DataProviderUtils;
-import com.linkedin.venice.utils.IntegrationTestPushUtils;
-import com.linkedin.venice.utils.PropertyBuilder;
-import com.linkedin.venice.utils.TestUtils;
-import com.linkedin.venice.utils.TestWriteUtils;
-import com.linkedin.venice.utils.Utils;
-import com.linkedin.venice.utils.VeniceProperties;
-import com.linkedin.venice.writer.update.UpdateBuilder;
-import com.linkedin.venice.writer.update.UpdateBuilderImpl;
-import io.tehuti.metrics.MetricsRepository;
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Map;
-import java.util.Properties;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Consumer;
-import org.apache.avro.Schema;
-import org.apache.avro.generic.GenericData;
-import org.apache.avro.generic.GenericDatumWriter;
-import org.apache.avro.generic.GenericRecord;
-import org.apache.avro.io.BinaryEncoder;
-import org.apache.avro.io.DatumWriter;
-import org.apache.avro.util.Utf8;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.testng.Assert;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
-
-public class TestRepush extends AbstractMultiRegionTest {
-  private static final Logger LOGGER = LogManager.getLogger(TestRepush.class);
-  private static final int TEST_TIMEOUT_MS = 180_000;
-  private static final int ASSERTION_TIMEOUT_MS = 60_000;
-
-  private String[] dcNames;
-
-  @Override
-  protected boolean shouldCreateD2Client() {
-    return true;
-  }
-
-  @Override
-  @BeforeClass(alwaysRun = true)
-  public void setUp() {
-    super.setUp();
-    dcNames = multiRegionMultiClusterWrapper.getChildRegionNames().toArray(new String[0]);
-  }
-
-  @Test(timeOut = TEST_TIMEOUT_MS, dataProvider = "True-and-False", dataProviderClass = DataProviderUtils.class)
-  public void testRepushWithChunkingFlagChanged(boolean useSpark) {
-    final String storeName = Utils.getUniqueString("reproduce");
-    String parentControllerUrl = getParentControllerUrl();
-    Schema keySchema = AvroCompatibilityHelper.parse(loadFileAsString("UserKey.avsc"));
-    Schema valueSchema = AvroCompatibilityHelper.parse(loadFileAsString("UserValue.avsc"));
-    Schema writeComputeSchema = WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(valueSchema);
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient.createNewStore(storeName, "test_owner", keySchema.toString(), valueSchema.toString()));
-      UpdateStoreQueryParams updateStoreParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setCompressionStrategy(CompressionStrategy.NO_OP)
-              .setWriteComputationEnabled(true)
-              .setHybridRewindSeconds(86400L)
-              .setHybridOffsetLagThreshold(10L);
-      ControllerResponse updateStoreResponse =
-          parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, updateStoreParams));
-      assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      assertFalse(response.isError(), "Empty push to parent colo should succeed");
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          60,
-          TimeUnit.SECONDS);
-
-      VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-
-      GenericRecord keyRecord = new GenericData.Record(keySchema);
-      keyRecord.put("learnerUrn", "urn:li:member:682787898");
-      keyRecord.put("query", "python");
-      GenericRecord checkpointKeyRecord = new GenericData.Record(keySchema);
-      checkpointKeyRecord.put("learnerUrn", "urn:li:member:123");
-      checkpointKeyRecord.put("query", "python");
-
-      GenericRecord partialUpdateRecord = new UpdateBuilderImpl(writeComputeSchema)
-          .setElementsToAddToListField("blockedContentsUrns", Collections.singletonList("urn:li:lyndaCourse:751323"))
-          .build();
-      try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
-        sendStreamingRecord(veniceProducer, storeName, keyRecord, partialUpdateRecord);
-      }
-
-      // Perform one time repush to make sure repush can handle RMD chunks data correctly.
-      Properties props =
-          IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, "dummyInputPath", storeName);
-      props.setProperty(SOURCE_KAFKA, "true");
-      props.setProperty(VENICE_REPUSH_SOURCE_PUBSUB_BROKER, veniceCluster.getPubSubBrokerWrapper().getAddress());
-      props.setProperty(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5");
-      if (useSpark) {
-        props.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-      }
-      IntegrationTestPushUtils.runVPJ(props);
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 2),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            GenericRecord value = (GenericRecord) storeReader.get(keyRecord).get();
-            assertNotNull(value, "key " + keyRecord + " should not be missing!");
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-
-        // Enable chunking
-        UpdateStoreQueryParams newUpdateStoreParams =
-            new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA).setChunkingEnabled(true);
-        updateStoreResponse =
-            parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, newUpdateStoreParams));
-        assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-        // Perform one time repush to make sure repush can handle chunks data correctly.
-        // intentionally stop re-consuming from RT so stale records don't affect the testing results
-        props.put(REWIND_TIME_IN_SECONDS_OVERRIDE, 0);
-        IntegrationTestPushUtils.runVPJ(props);
-
-        TestUtils.waitForNonDeterministicPushCompletion(
-            Version.composeKafkaTopic(storeName, 3),
-            parentControllerClient,
-            30,
-            TimeUnit.SECONDS);
-
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            GenericRecord value = (GenericRecord) storeReader.get(keyRecord).get();
-            assertNotNull(value, "key " + keyRecord + " should not be missing!");
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-        try (
-            VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
-          partialUpdateRecord = new UpdateBuilderImpl(writeComputeSchema)
-              .setElementsToAddToListField("blockedContentsUrns", Collections.singletonList("urn:li:lyndaCourse:1"))
-              .build();
-          sendStreamingRecord(veniceProducer, storeName, keyRecord, partialUpdateRecord);
-          partialUpdateRecord = new UpdateBuilderImpl(writeComputeSchema)
-              .setElementsToAddToListField("blockedContentsUrns", Collections.singletonList("urn:li:lyndaCourse:2"))
-              .build();
-          sendStreamingRecord(veniceProducer, storeName, keyRecord, partialUpdateRecord);
-          sendStreamingRecord(veniceProducer, storeName, checkpointKeyRecord, partialUpdateRecord);
-        }
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            GenericRecord value = (GenericRecord) storeReader.get(checkpointKeyRecord).get();
-            assertNotNull(value, "key " + checkpointKeyRecord + " should not be missing!");
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-        TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, () -> {
-          try {
-            GenericRecord value = (GenericRecord) storeReader.get(keyRecord).get();
-            assertNotNull(value, "key " + keyRecord + " should not be missing!");
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT_MS, dataProvider = "Compression-Boolean", dataProviderClass = DataProviderUtils.class)
-  public void testRepushWithTTLWithActiveActivePartialUpdateStore(
-      CompressionStrategy compressionStrategy,
-      boolean useSpark) {
-    final String storeName = Utils.getUniqueString("ttlRepushAAWC");
-    String parentControllerUrl = getParentControllerUrl();
-    Schema valueSchema = AvroCompatibilityHelper.parse(loadFileAsString("CollectionRecordV1.avsc"));
-    Schema partialUpdateSchema = WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(valueSchema);
-
-    Schema rmdSchema = RmdSchemaGenerator.generateMetadataSchema(valueSchema);
-    ReadOnlySchemaRepository schemaRepo = mock(ReadOnlySchemaRepository.class);
-    when(schemaRepo.getReplicationMetadataSchema(storeName, 1, 1)).thenReturn(new RmdSchemaEntry(1, 1, rmdSchema));
-    when(schemaRepo.getDerivedSchema(storeName, 1, 1)).thenReturn(new DerivedSchemaEntry(1, 1, partialUpdateSchema));
-    when(schemaRepo.getValueSchema(storeName, 1)).thenReturn(new SchemaEntry(1, valueSchema));
-    StringAnnotatedStoreSchemaCache stringAnnotatedStoreSchemaCache =
-        new StringAnnotatedStoreSchemaCache(storeName, schemaRepo);
-    RmdSerDe rmdSerDe = new RmdSerDe(stringAnnotatedStoreSchemaCache, 1);
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient
-              .createNewStore(storeName, "test_owner", STRING_SCHEMA.toString(), valueSchema.toString()));
-      UpdateStoreQueryParams updateStoreParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setPartitionCount(1)
-              .setCompressionStrategy(compressionStrategy)
-              .setWriteComputationEnabled(true)
-              .setActiveActiveReplicationEnabled(true)
-              .setChunkingEnabled(true)
-              .setRmdChunkingEnabled(true)
-              .setHybridRewindSeconds(1L)
-              .setHybridOffsetLagThreshold(1L);
-      ControllerResponse updateStoreResponse =
-          parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, updateStoreParams));
-      assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      assertFalse(response.isError(), "Empty push to parent colo should succeed");
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          60,
-          TimeUnit.SECONDS);
-    }
-
-    VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-
-    long STALE_TS = 99999L;
-    long FRESH_TS = 100000L;
-    String STRING_MAP_FILED = "stringMap";
-    String REGULAR_FIELD = "name";
-    String key1 = "key1";
-    String key2 = "key2";
-    String key3 = "key3";
-    String key4 = "key4";
-    String key5 = "key5";
-
-    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
-      /**
-       * Case 1: The record is partially stale, TTL repush should only keep the part that's fresh.
-       */
-
-      // This update is expected to be carried into TTL repush.
-      UpdateBuilder updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-      updateBuilder.setEntriesToAddToMapField(STRING_MAP_FILED, Collections.singletonMap("k1", "v1"));
-      sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), FRESH_TS);
-      // This update is expected to be WIPED OUT after TTL repush.
-      updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-      updateBuilder.setNewFieldValue(REGULAR_FIELD, "new_name");
-      updateBuilder.setEntriesToAddToMapField(STRING_MAP_FILED, Collections.singletonMap("k2", "v2"));
-      sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), STALE_TS);
-
-      /**
-       * Case 2: The record is fully stale, TTL repush should drop the record.
-       */
-
-      // This update is expected to be WIPED OUT after TTL repush.
-      updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-      updateBuilder.setNewFieldValue(REGULAR_FIELD, "new_name_2");
-      sendStreamingRecord(veniceProducer, storeName, key2, updateBuilder.build(), STALE_TS);
-
-      /**
-       * Case 3: The record is fully fresh, TTL repush should keep the record.
-       */
-
-      // This update is expected to be carried into TTL repush.
-      updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-      updateBuilder.setNewFieldValue(REGULAR_FIELD, "new_name_3");
-      sendStreamingRecord(veniceProducer, storeName, key3, updateBuilder.build(), FRESH_TS);
-
-      /**
-       * Case 4: The delete record is fully fresh, TTL repush should NOT FAIL due to this record.
-       */
-
-      sendStreamingRecord(veniceProducer, storeName, key4, updateBuilder.build(), FRESH_TS);
-      sendStreamingDeleteRecord(veniceProducer, storeName, key4, FRESH_TS + 1);
-
-      /**
-       * Case 5: The delete record is stale, TTL repush should keep new update.
-       */
-
-      sendStreamingDeleteRecord(veniceProducer, storeName, key5, STALE_TS);
-      updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-      updateBuilder.setNewFieldValue(REGULAR_FIELD, "new_name_5");
-      sendStreamingRecord(veniceProducer, storeName, key5, updateBuilder.build(), FRESH_TS);
-
-    }
-
-    /**
-     * Validate the data is ready in storage before TTL repush.
-     */
-    try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
-      TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-        try {
-          GenericRecord valueRecord = readValue(storeReader, key1);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name"));
-          assertNotNull(valueRecord.get(STRING_MAP_FILED));
-          Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-          assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          assertEquals(stringMapValue.get(new Utf8("k2")), new Utf8("v2"));
-
-          valueRecord = readValue(storeReader, key2);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_2"));
-
-          valueRecord = readValue(storeReader, key3);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_3"));
-
-          valueRecord = readValue(storeReader, key4);
-          assertNull(valueRecord);
-
-          valueRecord = readValue(storeReader, key5);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_5"));
-        } catch (Exception e) {
-          throw new VeniceException(e);
-        }
-      });
-    }
-
-    String kafkaTopic_v1 = Version.composeKafkaTopic(storeName, 1);
-    // Validate RMD bytes after initial update requests.
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key1, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      GenericRecord stringMapTsRecord = (GenericRecord) timestampRecord.get(STRING_MAP_FILED);
-      Assert.assertEquals(stringMapTsRecord.get(TOP_LEVEL_TS_FIELD_NAME), 0L);
-      Assert.assertEquals(stringMapTsRecord.get(ACTIVE_ELEM_TS_FIELD_NAME), Arrays.asList(STALE_TS, FRESH_TS));
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), STALE_TS);
-    });
-
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key2, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), STALE_TS);
-    });
-
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key3, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), FRESH_TS);
-    });
-
-    // Perform one time repush to make sure repush can handle RMD chunks data correctly.
-    Properties props =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, "dummyInputPath", storeName);
-    props.setProperty(SOURCE_KAFKA, "true");
-    props.setProperty(VENICE_REPUSH_SOURCE_PUBSUB_BROKER, veniceCluster.getPubSubBrokerWrapper().getAddress());
-    props.setProperty(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5");
-    props.setProperty(REPUSH_TTL_ENABLE, "true");
-    // Override the TTL repush start TS to work with logical TS setup.
-    props.setProperty(REPUSH_TTL_START_TIMESTAMP, String.valueOf(FRESH_TS));
-    // Override the rewind time to make sure not to consume 24hrs data from RT topic.
-    props.put(REWIND_TIME_IN_SECONDS_OVERRIDE, 0);
-    if (useSpark) {
-      props.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    }
-    IntegrationTestPushUtils.runVPJ(props);
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 2),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-    }
-
-    /**
-     * Validate the data is ready in storage after TTL repush.
-     */
-    try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
-      TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-        try {
-          // Key 1 is partially preserved.
-          GenericRecord valueRecord = readValue(storeReader, key1);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("default_name"));
-          assertNotNull(valueRecord.get(STRING_MAP_FILED));
-          Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-          assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          assertNull(stringMapValue.get(new Utf8("k2")));
-
-          // Key 2 is fully removed.
-          valueRecord = readValue(storeReader, key2);
-          assertNull(valueRecord);
-
-          // Key 3 is fully preserved.
-          valueRecord = readValue(storeReader, key3);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_3"));
-
-          valueRecord = readValue(storeReader, key4);
-          assertNull(valueRecord);
-
-          valueRecord = readValue(storeReader, key5);
-          assertNotNull(valueRecord);
-          assertEquals(valueRecord.get(REGULAR_FIELD), new Utf8("new_name_5"));
-        } catch (Exception e) {
-          throw new VeniceException(e);
-        }
-      });
-    }
-
-    String kafkaTopic_v2 = Version.composeKafkaTopic(storeName, 2);
-    // Validate RMD bytes after TTL repush.
-    validateRmdData(rmdSerDe, kafkaTopic_v2, key1, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      GenericRecord stringMapTsRecord = (GenericRecord) timestampRecord.get(STRING_MAP_FILED);
-      Assert.assertEquals(stringMapTsRecord.get(TOP_LEVEL_TS_FIELD_NAME), STALE_TS);
-      Assert.assertEquals(stringMapTsRecord.get(ACTIVE_ELEM_TS_FIELD_NAME), Collections.singletonList(FRESH_TS));
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), STALE_TS);
-    });
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key2, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), STALE_TS);
-    });
-
-    validateRmdData(rmdSerDe, kafkaTopic_v1, key3, rmdWithValueSchemaId -> {
-      GenericRecord timestampRecord = (GenericRecord) rmdWithValueSchemaId.getRmdRecord().get(TIMESTAMP_FIELD_NAME);
-      Assert.assertEquals(timestampRecord.get(REGULAR_FIELD), FRESH_TS);
-    });
-  }
-
-  @Test(timeOut = TEST_TIMEOUT_MS, dataProvider = "True-and-False", dataProviderClass = DataProviderUtils.class)
-  public void testRepushWithDeleteRecord(boolean useSpark) {
-    final String storeName = Utils.getUniqueString("testRepushWithDeleteRecord");
-    String parentControllerUrl = getParentControllerUrl();
-    Schema valueSchema = AvroCompatibilityHelper.parse(loadFileAsString("CollectionRecordV1.avsc"));
-    Schema partialUpdateSchema = WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(valueSchema);
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient
-              .createNewStore(storeName, "test_owner", STRING_SCHEMA.toString(), valueSchema.toString()));
-      UpdateStoreQueryParams updateStoreParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setPartitionCount(1)
-              .setCompressionStrategy(CompressionStrategy.NO_OP)
-              .setWriteComputationEnabled(true)
-              .setActiveActiveReplicationEnabled(true)
-              .setChunkingEnabled(true)
-              .setRmdChunkingEnabled(true)
-              .setHybridRewindSeconds(1L)
-              .setHybridOffsetLagThreshold(1L);
-      ControllerResponse updateStoreResponse =
-          parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, updateStoreParams));
-      assertFalse(updateStoreResponse.isError(), "Update store got error: " + updateStoreResponse.getError());
-
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      assertFalse(response.isError(), "Empty push to parent colo should succeed");
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          60,
-          TimeUnit.SECONDS);
-    }
-
-    VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    String STRING_MAP_FILED = "stringMap";
-    String key1 = "key1";
-    String key2 = "key2";
-    String key3 = "key3";
-
-    UpdateBuilder updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-    updateBuilder.setEntriesToAddToMapField(STRING_MAP_FILED, Collections.singletonMap("k1", "v1"));
-    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
-      sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), 10000L);
-      sendStreamingDeleteRecord(veniceProducer, storeName, key1, 10001L);
-      sendStreamingRecord(veniceProducer, storeName, key2, updateBuilder.build(), 10001L);
-    }
-
-    Properties props =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, "dummyInputPath", storeName);
-    props.setProperty(SOURCE_KAFKA, "true");
-    props.setProperty(VENICE_REPUSH_SOURCE_PUBSUB_BROKER, veniceCluster.getPubSubBrokerWrapper().getAddress());
-    props.setProperty(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5");
-    if (useSpark) {
-      props.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    }
-    IntegrationTestPushUtils.runVPJ(props);
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 2),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-    }
-    int numberOfChildDatacenters = childDatacenters.size();
-    for (int i = 0; i < numberOfChildDatacenters; i++) {
-      VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(cluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-          try {
-            // Key 1 is deleted.
-            GenericRecord valueRecord = readValue(storeReader, key1);
-            assertNull(valueRecord);
-
-            // Key 2 is preserved.
-            valueRecord = readValue(storeReader, key2);
-            assertNotNull(valueRecord);
-            assertNotNull(valueRecord.get(STRING_MAP_FILED));
-            Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-            assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
-
-    String baseDataPath = Utils.getTempDataDirectory().getAbsolutePath();
-    VeniceProperties backendConfig = new PropertyBuilder().put(CLIENT_USE_SYSTEM_STORE_REPOSITORY, true)
-        .put(CLIENT_SYSTEM_STORE_REPOSITORY_REFRESH_INTERVAL_SECONDS, 1)
-        .put(DATA_BASE_PATH, baseDataPath)
-        .put(PERSISTENCE_TYPE, ROCKS_DB)
-        .put(ROCKSDB_PLAIN_TABLE_FORMAT_ENABLED, "false")
-        .put(ROCKSDB_BLOCK_CACHE_SIZE_IN_BYTES, 2 * 1024 * 1024L)
-        .put(DAVINCI_PUSH_STATUS_CHECK_INTERVAL_IN_MS, 1000)
-        .build();
-
-    MetricsRepository metricsRepository = new VeniceMetricsRepository();
-    try (CachingDaVinciClientFactory factory = new CachingDaVinciClientFactory(
-        d2ClientDC0,
-        VeniceRouterWrapper.CLUSTER_DISCOVERY_D2_SERVICE_NAME,
-        metricsRepository,
-        backendConfig)) {
-      DaVinciClient<Integer, Object> client =
-          factory.getAndStartGenericAvroClient(storeName, new DaVinciConfig().setStorageClass(StorageClass.DISK));
-      client.subscribeAll().get();
-    } catch (Exception e) {
-      throw new VeniceException(e);
-    }
-
-    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
-      sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), 9999L);
-      sendStreamingRecord(veniceProducer, storeName, key3, updateBuilder.build(), 10000L);
-    }
-
-    for (int i = 0; i < numberOfChildDatacenters; i++) {
-      VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(cluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-          try {
-            // Key 1 is deleted.
-            GenericRecord valueRecord = readValue(storeReader, key1);
-            assertNull(valueRecord);
-
-            // Key 3 is preserved.
-            valueRecord = readValue(storeReader, key3);
-            assertNotNull(valueRecord);
-            assertNotNull(valueRecord.get(STRING_MAP_FILED));
-            Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-            assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
-
-    try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
-      sendStreamingRecord(veniceProducer, storeName, key1, updateBuilder.build(), 12345L);
-    }
-
-    for (int i = 0; i < numberOfChildDatacenters; i++) {
-      VeniceClusterWrapper cluster = childDatacenters.get(i).getClusters().get(CLUSTER_NAME);
-      try (AvroGenericStoreClient<Object, Object> storeReader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(cluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, true, () -> {
-          try {
-            // Key 1 is preserved.
-            GenericRecord valueRecord = readValue(storeReader, key1);
-            assertNotNull(valueRecord);
-            assertNotNull(valueRecord.get(STRING_MAP_FILED));
-            Map<Utf8, Utf8> stringMapValue = (Map<Utf8, Utf8>) valueRecord.get(STRING_MAP_FILED);
-            assertEquals(stringMapValue.get(new Utf8("k1")), new Utf8("v1"));
-          } catch (Exception e) {
-            throw new VeniceException(e);
-          }
-        });
-      }
-    }
-  }
-
-  /**
-   * Test that a batch-store repush correctly reads data from a cross-fabric input (dc-1) when
-   * the NR source is dc-0. Parameterized for both Spark and Hadoop MR compute engines.
-   *
-   * <p><b>Setup:</b> A batch-only store with NR enabled, AA enabled, NR source = dc-0.
-   * After a v1 batch push, we repush from dc-1 (KAFKA_INPUT_FABRIC = dc-1).
-   *
-   * <p><b>Key verification:</b>
-   * <ul>
-   *   <li>The VenicePushJob's repushSourcePubsubBroker should resolve to dc-1's broker, confirming
-   *       the consume fabric was correctly set.</li>
-   *   <li>All records should be present in both DCs after repush, confirming that data was
-   *       correctly produced to the NR source (dc-0) and replicated.</li>
-   * </ul>
-   *
-   * <p><b>Expected outcome:</b> Both MR and Spark variants should pass.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS, dataProvider = "True-and-False", dataProviderClass = DataProviderUtils.class)
-  public void testBatchStoreRepushWithCrossFabricInput(boolean useSpark) throws Exception {
-    String engine = useSpark ? "Spark" : "MR";
-    String storeName = Utils.getUniqueString("batch-repush-xfabric-" + engine.toLowerCase());
-    String parentControllerUrls = multiRegionMultiClusterWrapper.getControllerConnectString();
-
-    /*
-     * Step 1: Write 50 simple avro records to a temp directory.
-     * Keys are "1".."50", values are "test_name_1".."test_name_50".
-     */
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema(inputDir, 50);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String valueSchemaStr = recordSchema.getField(DEFAULT_VALUE_FIELD_PROP).schema().toString();
-
-    VeniceMultiClusterWrapper dc0 = childDatacenters.get(0);
-    VeniceMultiClusterWrapper dc1 = childDatacenters.get(1);
-    String dc0KafkaUrl = dc0.getPubSubBrokerWrapper().getAddress();
-    String dc1KafkaUrl = dc1.getPubSubBrokerWrapper().getAddress();
-    assertNotEquals(dc0KafkaUrl, dc1KafkaUrl, "DCs must have different Kafka brokers for this test");
-
-    /*
-     * Step 2: Create a batch-only store with NR enabled, AA enabled, NR source = dc-0.
-     * No hybrid config is set — this is a pure batch store.
-     */
-    Properties batchProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    batchProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    createStoreForJob(
-        CLUSTER_NAME,
-        keySchemaStr,
-        valueSchemaStr,
-        batchProps,
-        new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-            .setPartitionCount(2)
-            .setNativeReplicationEnabled(true)
-            .setActiveActiveReplicationEnabled(true)
-            .setNativeReplicationSourceFabric(dcNames[0])).close();
-
-    /*
-     * Step 3: Batch push v1 and verify data is present in both DCs.
-     */
-    try (VenicePushJob batchPush = new VenicePushJob("batch-push-v1-" + engine, batchProps)) {
-      batchPush.run();
-    }
-    try (ControllerClient parentClient = new ControllerClient(CLUSTER_NAME, parentControllerUrls)) {
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-        for (int version: parentClient.getStore(storeName).getStore().getColoToCurrentVersions().values()) {
-          assertEquals(version, 1, "All DCs should be on v1 after batch push");
-        }
-      });
-    }
-    verifyBatchData(storeName, 50, 0);
-    verifyBatchData(storeName, 50, 1);
-
-    /*
-     * Step 4: Repush from dc-1 (cross-fabric). The controller should resolve KAFKA_INPUT_FABRIC
-     * to dc-1's broker URL. When useSpark=true, DataWriterSparkJob is used.
-     *
-     * We use a direct VenicePushJob (not runVPJ) so we can inspect getRepushSourcePubsubBroker()
-     * to verify the consume-fabric was correctly resolved to dc-1.
-     */
-    Properties repushProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    repushProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    repushProps.put(SOURCE_KAFKA, "true");
-    repushProps.put(KAFKA_INPUT_FABRIC, dcNames[1]);
-    repushProps.put(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5000");
-    if (useSpark) {
-      repushProps.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    }
-
-    try (VenicePushJob repushJob = new VenicePushJob(engine + "-repush-from-dc1", repushProps)) {
-      repushJob.run();
-
-      /*
-       * Step 5: Verify consume fabric. The repushSourcePubsubBroker should point to dc-1's
-       * broker, confirming that the repush reads data from dc-1 (not dc-0).
-       */
-      assertEquals(
-          repushJob.getRepushSourcePubsubBroker(),
-          dc1KafkaUrl,
-          "Repush should consume from dc-1 (KAFKA_INPUT_FABRIC=" + dcNames[1] + "), "
-              + "but repushSourcePubsubBroker points elsewhere");
-      if (useSpark) {
-        repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-          assertStage(snapshot, "compaction", 50, 50);
-          assertSinkStage(snapshot, "kafka_write", 50);
-          assertNull(snapshot.getStage("chunk_assembly"));
-          assertNull(snapshot.getStage("compression_reencode"));
-          assertNull(snapshot.getStage("ttl_filter"));
-          LOGGER.info("Cross-fabric batch repush metrics:\n{}", snapshot.getFormattedReport());
-        });
-      }
-    }
-
-    /*
-     * Step 6: Wait for v2 to become current in all DCs and verify data.
-     * If the produce fabric is correct (NR source = dc-0), all records should be
-     * present in both DCs after replication.
-     */
-    try (ControllerClient parentClient = new ControllerClient(CLUSTER_NAME, parentControllerUrls)) {
-      TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, () -> {
-        for (int version: parentClient.getStore(storeName).getStore().getColoToCurrentVersions().values()) {
-          assertEquals(version, 2, "All DCs should be on v2 after repush");
-        }
-      });
-    }
-    verifyBatchData(storeName, 50, 0);
-    verifyBatchData(storeName, 50, 1);
-    LOGGER.info("[{}] testBatchStoreRepushWithCrossFabricInput passed", engine);
-  }
-
-  /**
-   * Test that a hybrid-store repush correctly reads data from a cross-fabric input (dc-1)
-   * by verifying at the data level that only dc-1's data appears in the repushed version.
-   * Parameterized for both Spark and Hadoop MR compute engines.
-   *
-   * <p><b>Test design:</b> The store is created with NR enabled but AA disabled. After batch
-   * push v1, we send new streaming records ONLY to dc-0. Without AA, these records do NOT
-   * replicate to dc-1. We then repush from dc-1 with REWIND_TIME_IN_SECONDS_OVERRIDE=0
-   * (no RT consumption). If the repush correctly reads from dc-1, the dc-0-only streaming
-   * records should be ABSENT from v2 — proving the data was sourced from dc-1.
-   *
-   * <p><b>Expected outcome:</b> Both MR and Spark variants should pass.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS, dataProvider = "True-and-False", dataProviderClass = DataProviderUtils.class)
-  public void testHybridStoreRepushWithCrossFabricInput(boolean useSpark) throws Exception {
-    String engine = useSpark ? "Spark" : "MR";
-    String storeName = Utils.getUniqueString("hybrid-repush-xfabric-" + engine.toLowerCase());
-    String parentControllerUrls = multiRegionMultiClusterWrapper.getControllerConnectString();
-
-    /*
-     * Step 1: Write 50 simple avro records to a temp directory.
-     * Keys are "1".."50", values are "test_name_1".."test_name_50".
-     */
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = TestWriteUtils.writeSimpleAvroFileWithStringToStringSchema(inputDir, 50);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String keySchemaStr = recordSchema.getField(DEFAULT_KEY_FIELD_PROP).schema().toString();
-    String valueSchemaStr = recordSchema.getField(DEFAULT_VALUE_FIELD_PROP).schema().toString();
-
-    VeniceMultiClusterWrapper dc0 = childDatacenters.get(0);
-    VeniceMultiClusterWrapper dc1 = childDatacenters.get(1);
-    String dc0KafkaUrl = dc0.getPubSubBrokerWrapper().getAddress();
-    String dc1KafkaUrl = dc1.getPubSubBrokerWrapper().getAddress();
-    assertNotEquals(dc0KafkaUrl, dc1KafkaUrl, "DCs must have different Kafka brokers for this test");
-
-    /*
-     * Step 2: Create a hybrid store with NR enabled but AA DISABLED, NR source = dc-0.
-     * AA is intentionally off so that streaming records sent to dc-0 do NOT replicate
-     * to dc-1, creating the data asymmetry needed to verify consume-fabric correctness.
-     * hybridRewindSeconds and hybridOffsetLagThreshold are set to make it a hybrid store.
-     */
-    Properties batchProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    batchProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    createStoreForJob(
-        CLUSTER_NAME,
-        keySchemaStr,
-        valueSchemaStr,
-        batchProps,
-        new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-            .setPartitionCount(2)
-            .setNativeReplicationEnabled(true)
-            .setNativeReplicationSourceFabric(dcNames[0])
-            .setHybridRewindSeconds(86400L)
-            .setHybridOffsetLagThreshold(10L)).close();
-
-    /*
-     * Step 3: Batch push v1 and verify data is present in both DCs.
-     */
-    try (VenicePushJob batchPush = new VenicePushJob("hybrid-batch-push-v1-" + engine, batchProps)) {
-      batchPush.run();
-    }
-    try (ControllerClient parentClient = new ControllerClient(CLUSTER_NAME, parentControllerUrls)) {
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-        for (int version: parentClient.getStore(storeName).getStore().getColoToCurrentVersions().values()) {
-          assertEquals(version, 1, "All DCs should be on v1 after batch push");
-        }
-      });
-    }
-    verifyBatchData(storeName, 50, 0);
-    verifyBatchData(storeName, 50, 1);
-
-    /*
-     * Step 4: Send new streaming records ONLY to dc-0. These records use keys "51".."60"
-     * which are NOT in the original batch data (keys "1".."50"). Since AA is disabled
-     * (the store was created without it), these records will exist only in dc-0's
-     * real-time topic and will be consumed only by dc-0's storage nodes.
-     */
-    VeniceClusterWrapper dc0Cluster = dc0.getClusters().get(CLUSTER_NAME);
-    try (VeniceSystemProducer dc0Producer = getSamzaProducer(dc0Cluster, storeName, Version.PushType.STREAM)) {
-      for (int i = 51; i <= 60; i++) {
-        sendStreamingRecord(dc0Producer, storeName, Integer.toString(i), "stream_value_" + i);
-      }
-    }
-
-    /*
-     * Step 5: Wait until dc-0 has the new streaming records. Query dc-0's router to confirm.
-     */
-    VeniceClusterWrapper dc0ClusterForRead = dc0.getClusters().get(CLUSTER_NAME);
-    try (AvroGenericStoreClient<String, Object> dc0Reader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(dc0ClusterForRead.getRandomRouterURL()))) {
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-        Object val = dc0Reader.get("55").get();
-        assertNotNull(val, "dc-0 should have streaming record for key 55");
-      });
-    }
-
-    /*
-     * Step 6: Confirm dc-1 does NOT have the new streaming records.
-     * Since AA is disabled, the records sent to dc-0 should not have replicated to dc-1.
-     */
-    VeniceClusterWrapper dc1Cluster = dc1.getClusters().get(CLUSTER_NAME);
-    try (AvroGenericStoreClient<String, Object> dc1Reader = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(dc1Cluster.getRandomRouterURL()))) {
-      for (int i = 51; i <= 60; i++) {
-        Object val = dc1Reader.get(Integer.toString(i)).get();
-        assertNull(val, "dc-1 should NOT have streaming record for key " + i + " (AA is disabled)");
-      }
-    }
-
-    /*
-     * Step 7: Repush from dc-1 — the colo WITHOUT the new streaming records.
-     * REWIND_TIME_IN_SECONDS_OVERRIDE=0 ensures no RT consumption, so the repush only gets
-     * VT (Version Topic) data from dc-1. If the repush correctly reads from dc-1, the
-     * dc-0-only streaming records (keys 51-60) should NOT appear in v2.
-     */
-    Properties repushProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    repushProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    repushProps.put(SOURCE_KAFKA, "true");
-    repushProps.put(KAFKA_INPUT_FABRIC, dcNames[1]);
-    repushProps.put(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5000");
-    repushProps.put(REWIND_TIME_IN_SECONDS_OVERRIDE, 0);
-    if (useSpark) {
-      repushProps.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    }
-
-    try (VenicePushJob repushJob = new VenicePushJob(engine + "-hybrid-repush-from-dc1", repushProps)) {
-      repushJob.run();
-
-      /*
-       * Step 8: Verify consume fabric. The repushSourcePubsubBroker should point to dc-1's
-       * broker, confirming that the repush reads data from dc-1.
-       */
-      assertEquals(
-          repushJob.getRepushSourcePubsubBroker(),
-          dc1KafkaUrl,
-          "Repush should consume from dc-1 (KAFKA_INPUT_FABRIC=" + dcNames[1] + "), "
-              + "but repushSourcePubsubBroker points elsewhere");
-      if (useSpark) {
-        repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-          assertStage(snapshot, "compaction", 50, 50);
-          assertSinkStage(snapshot, "kafka_write", 50);
-          assertNull(snapshot.getStage("chunk_assembly"));
-          assertNull(snapshot.getStage("compression_reencode"));
-          assertNull(snapshot.getStage("ttl_filter"));
-          LOGGER.info("Cross-fabric hybrid repush metrics:\n{}", snapshot.getFormattedReport());
-        });
-      }
-    }
-
-    /*
-     * Step 9: Wait for v2 to become current in all DCs.
-     */
-    try (ControllerClient parentClient = new ControllerClient(CLUSTER_NAME, parentControllerUrls)) {
-      TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, () -> {
-        for (int version: parentClient.getStore(storeName).getStore().getColoToCurrentVersions().values()) {
-          assertEquals(version, 2, "All DCs should be on v2 after repush");
-        }
-      });
-    }
-
-    /*
-     * Step 10: Verify the dc-0-only streaming records (keys 51-60) are ABSENT from v2.
-     * This proves the repush read from dc-1 (which doesn't have them), not from dc-0.
-     * We check both DCs to be thorough.
-     */
-    for (int dcIndex = 0; dcIndex < childDatacenters.size(); dcIndex++) {
-      VeniceClusterWrapper cluster = childDatacenters.get(dcIndex).getClusters().get(CLUSTER_NAME);
-      try (AvroGenericStoreClient<String, Object> reader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(cluster.getRandomRouterURL()))) {
-        int dc = dcIndex;
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-          for (int i = 51; i <= 60; i++) {
-            Object val = reader.get(Integer.toString(i)).get();
-            assertNull(
-                val,
-                "Key " + i + " (dc-0-only streaming record) should be ABSENT in dc-" + dc
-                    + " after repush from dc-1. Its presence would mean the repush read from "
-                    + "dc-0 instead of dc-1.");
-          }
-        });
-      }
-    }
-
-    /*
-     * Step 11: Verify the original batch records (keys 1-50) ARE present in both DCs.
-     * This confirms the repush successfully transferred the batch data.
-     */
-    verifyBatchData(storeName, 50, 0);
-    verifyBatchData(storeName, 50, 1);
-    LOGGER.info("[{}] testHybridStoreRepushWithCrossFabricInput passed", engine);
-  }
-
-  /**
-   * Baseline KIF repush: batch push then repush twice (combiner=true, combiner=false).
-   * Exercises: compaction + kafka_write stages.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS)
-  public void testKafkaInputBatchJob() throws Exception {
-    String storeName = Utils.getUniqueString("repush-basic");
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String keySchemaStr = recordSchema.getField("key").schema().toString();
-    String valueSchemaStr = recordSchema.getField("value").schema().toString();
-
-    Properties batchProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    batchProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    batchProps.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    createStoreForJob(
-        CLUSTER_NAME,
-        keySchemaStr,
-        valueSchemaStr,
-        batchProps,
-        new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-            .setPartitionCount(2)
-            .setNativeReplicationEnabled(true)
-            .setActiveActiveReplicationEnabled(true)
-            .setNativeReplicationSourceFabric(dcNames[0])).close();
-
-    try (VenicePushJob batchPush = new VenicePushJob("repush-basic-batch-v1", batchProps)) {
-      batchPush.run();
-    }
-    waitForVersion(storeName, 1);
-
-    // Repush twice: combiner=true then combiner=false
-    for (String combiner: new String[] { "true", "false" }) {
-      Properties repushProps = buildRepushProps(storeName, inputDirPath);
-      repushProps.setProperty(KAFKA_INPUT_COMBINER_ENABLED, combiner);
-      try (VenicePushJob repushJob = new VenicePushJob("repush-basic-combiner-" + combiner, repushProps)) {
-        repushJob.run();
-        repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-          assertStage(snapshot, "compaction", 100, 100);
-          assertSinkStage(snapshot, "kafka_write", 100);
-          assertNull(snapshot.getStage("chunk_assembly"), "chunk_assembly should not be registered");
-          assertNull(snapshot.getStage("compression_reencode"), "compression_reencode should not be registered");
-          assertNull(snapshot.getStage("ttl_filter"), "ttl_filter should not be registered");
-          LOGGER.info("Repush (combiner={}) metrics:\n{}", combiner, snapshot.getFormattedReport());
-        });
-      }
-    }
-    verifyBatchData(storeName, 100, 0);
-  }
-
-  /**
-   * KIF repush with GZIP compression.
-   * Exercises: compaction + kafka_write stages.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS)
-  public void testCompressingRecordRepush() throws Exception {
-    String storeName = Utils.getUniqueString("repush-gzip");
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String keySchemaStr = recordSchema.getField("key").schema().toString();
-    String valueSchemaStr = recordSchema.getField("value").schema().toString();
-
-    Properties batchProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    batchProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    batchProps.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    createStoreForJob(
-        CLUSTER_NAME,
-        keySchemaStr,
-        valueSchemaStr,
-        batchProps,
-        new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-            .setPartitionCount(2)
-            .setNativeReplicationEnabled(true)
-            .setActiveActiveReplicationEnabled(true)
-            .setNativeReplicationSourceFabric(dcNames[0])
-            .setCompressionStrategy(CompressionStrategy.GZIP)).close();
-
-    try (VenicePushJob batchPush = new VenicePushJob("repush-gzip-batch-v1", batchProps)) {
-      batchPush.run();
-    }
-    waitForVersion(storeName, 1);
-
-    // Repush
-    Properties repushProps = buildRepushProps(storeName, inputDirPath);
-    try (VenicePushJob repushJob = new VenicePushJob("repush-gzip-repush", repushProps)) {
-      repushJob.run();
-      repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-        assertStage(snapshot, "compaction", 100, 100);
-        assertSinkStage(snapshot, "kafka_write", 100);
-        assertNull(snapshot.getStage("chunk_assembly"));
-        assertNull(snapshot.getStage("compression_reencode"));
-        assertNull(snapshot.getStage("ttl_filter"));
-        LOGGER.info("GZIP repush metrics:\n{}", snapshot.getFormattedReport());
-      });
-    }
-    waitForVersion(storeName, 2);
-    verifyBatchData(storeName, 100, 0);
-  }
-
-  /**
-   * KIF repush with ZSTD_WITH_DICT compression. Verifies dictionary differs after repush.
-   * Exercises: compaction + compression_reencode + kafka_write stages.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS * 2)
-  public void testZstdCompressionRepush() throws Exception {
-    String storeName = Utils.getUniqueString("repush-zstd");
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String keySchemaStr = recordSchema.getField("key").schema().toString();
-    String valueSchemaStr = recordSchema.getField("value").schema().toString();
-
-    // Batch push v1 (no compression)
-    Properties batchProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    batchProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    batchProps.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    createStoreForJob(
-        CLUSTER_NAME,
-        keySchemaStr,
-        valueSchemaStr,
-        batchProps,
-        new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-            .setPartitionCount(3)
-            .setNativeReplicationEnabled(true)
-            .setActiveActiveReplicationEnabled(true)
-            .setNativeReplicationSourceFabric(dcNames[0])).close();
-
-    try (VenicePushJob batchPush = new VenicePushJob("repush-zstd-batch-v1", batchProps)) {
-      batchPush.run();
-    }
-    waitForVersion(storeName, 1);
-
-    // Enable ZSTD compression and push v2
-    try (ControllerClient controllerClient =
-        new ControllerClient(CLUSTER_NAME, multiRegionMultiClusterWrapper.getControllerConnectString())) {
-      controllerClient.updateStore(
-          storeName,
-          new UpdateStoreQueryParams().setCompressionStrategy(CompressionStrategy.ZSTD_WITH_DICT));
-    }
-
-    try (VenicePushJob batchPush2 = new VenicePushJob("repush-zstd-batch-v2", batchProps)) {
-      batchPush2.run();
-    }
-    waitForVersion(storeName, 2);
-
-    // Repush from v2 (ZSTD)
-    Properties repushProps = buildRepushProps(storeName, inputDirPath);
-    try (VenicePushJob repushJob = new VenicePushJob("repush-zstd-repush-v3", repushProps)) {
-      repushJob.run();
-      repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-        assertStage(snapshot, "compaction", 100, 100);
-        assertNotNull(snapshot.getStage("compression_reencode"), "compression_reencode should be registered");
-        StageSummary compress = snapshot.getStage("compression_reencode");
-        assertEquals(compress.getRecordsIn(), 100, "compression_reencode recordsIn");
-        assertEquals(compress.getRecordsOut(), 100, "compression_reencode recordsOut");
-        assertTrue(compress.getBytesIn() > 0, "compression_reencode bytesIn should be > 0");
-        assertTrue(compress.getBytesOut() > 0, "compression_reencode bytesOut should be > 0");
-        assertSinkStage(snapshot, "kafka_write", 100);
-        assertNull(snapshot.getStage("chunk_assembly"));
-        assertNull(snapshot.getStage("ttl_filter"));
-        LOGGER.info("ZSTD repush metrics:\n{}", snapshot.getFormattedReport());
-      });
-    }
-    waitForVersion(storeName, 3);
-    verifyBatchData(storeName, 100, 0);
-  }
-
-  /**
-   * KIF repush on AA hybrid store.
-   * Exercises: compaction + kafka_write stages.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS)
-  public void testKafkaInputAAStoreRepush() throws Exception {
-    String storeName = Utils.getUniqueString("repush-aa");
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = writeSimpleAvroFileWithStringToStringSchema(inputDir);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String keySchemaStr = recordSchema.getField("key").schema().toString();
-    String valueSchemaStr = recordSchema.getField("value").schema().toString();
-
-    Properties batchProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    batchProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    batchProps.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    createStoreForJob(
-        CLUSTER_NAME,
-        keySchemaStr,
-        valueSchemaStr,
-        batchProps,
-        new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-            .setPartitionCount(2)
-            .setNativeReplicationEnabled(true)
-            .setActiveActiveReplicationEnabled(true)
-            .setNativeReplicationSourceFabric(dcNames[0])
-            .setHybridRewindSeconds(5)
-            .setHybridOffsetLagThreshold(2)).close();
-
-    try (VenicePushJob batchPush = new VenicePushJob("repush-aa-batch-v1", batchProps)) {
-      batchPush.run();
-    }
-    waitForVersion(storeName, 1);
-
-    // Repush
-    Properties repushProps = buildRepushProps(storeName, inputDirPath);
-    try (VenicePushJob repushJob = new VenicePushJob("repush-aa-repush", repushProps)) {
-      repushJob.run();
-      repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-        assertStage(snapshot, "compaction", 100, 100);
-        assertSinkStage(snapshot, "kafka_write", 100);
-        assertNull(snapshot.getStage("chunk_assembly"));
-        assertNull(snapshot.getStage("compression_reencode"));
-        assertNull(snapshot.getStage("ttl_filter"));
-        LOGGER.info("AA repush metrics:\n{}", snapshot.getFormattedReport());
-      });
-    }
-    waitForVersion(storeName, 2);
-    verifyBatchData(storeName, 100, 0);
-  }
-
-  /**
-   * KIF repush with 1 record and 3 partitions. Tests that the first mapper gets only control
-   * messages and that maybeSprayAllPartitions is NOT invoked.
-   * Exercises: compaction + kafka_write stages.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS)
-  public void testReducerCountValidation() throws Exception {
-    String storeName = Utils.getUniqueString("repush-reducer-count");
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = writeSimpleAvroFileWithStringToStringSchema(inputDir, 1);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String keySchemaStr = recordSchema.getField("key").schema().toString();
-    String valueSchemaStr = recordSchema.getField("value").schema().toString();
-
-    Properties batchProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    batchProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    batchProps.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    createStoreForJob(
-        CLUSTER_NAME,
-        keySchemaStr,
-        valueSchemaStr,
-        batchProps,
-        new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-            .setPartitionCount(3)
-            .setNativeReplicationEnabled(true)
-            .setActiveActiveReplicationEnabled(true)
-            .setNativeReplicationSourceFabric(dcNames[0])).close();
-
-    try (VenicePushJob batchPush = new VenicePushJob("repush-reducer-count-batch-v1", batchProps)) {
-      batchPush.run();
-    }
-    waitForVersion(storeName, 1);
-
-    // Repush with KAFKA_INPUT_MAX_RECORDS_PER_MAPPER=2 to force empty first mapper
-    Properties repushProps = buildRepushProps(storeName, inputDirPath);
-    repushProps.put(KAFKA_INPUT_TOPIC, Version.composeKafkaTopic(storeName, 1));
-    repushProps.put(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "2");
-    VeniceClusterWrapper dc0Cluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    repushProps.setProperty(VENICE_REPUSH_SOURCE_PUBSUB_BROKER, dc0Cluster.getPubSubBrokerWrapper().getAddress());
-    try (VenicePushJob repushJob = new VenicePushJob("repush-reducer-count-repush", repushProps)) {
-      repushJob.run();
-      repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-        assertStage(snapshot, "compaction", 1, 1);
-        assertSinkStage(snapshot, "kafka_write", 1);
-        assertNull(snapshot.getStage("chunk_assembly"));
-        assertNull(snapshot.getStage("compression_reencode"));
-        assertNull(snapshot.getStage("ttl_filter"));
-        LOGGER.info("Reducer count repush metrics:\n{}", snapshot.getFormattedReport());
-      });
-    }
-    waitForVersion(storeName, 2);
-    verifyBatchData(storeName, 1, 0);
-  }
-
-  /**
-   * KIF repush with large (3MB) chunked values.
-   * Exercises: chunk_assembly + kafka_write stages.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS * 2, dataProvider = "True-and-False", dataProviderClass = DataProviderUtils.class)
-  public void testKafkaInputBatchJobWithLargeValues(boolean sendDirectControlMessage) throws Exception {
-    int largeValueSize = 3 * 1024 * 1024; // 3 MB — exceeds 950KB chunk threshold
-    int numberOfRecords = 5;
-    String storeName = Utils.getUniqueString("repush-large-values");
-    File inputDir = getTempDataDirectory();
-    Schema recordSchema = writeSimpleAvroFileWithCustomSize(inputDir, numberOfRecords, 0, largeValueSize);
-    String inputDirPath = "file:" + inputDir.getAbsolutePath();
-    String keySchemaStr = recordSchema.getField("key").schema().toString();
-    String valueSchemaStr = recordSchema.getField("value").schema().toString();
-
-    // Batch push v1 with chunking enabled
-    Properties batchProps =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    batchProps.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    batchProps.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    createStoreForJob(
-        CLUSTER_NAME,
-        keySchemaStr,
-        valueSchemaStr,
-        batchProps,
-        new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-            .setPartitionCount(2)
-            .setChunkingEnabled(true)
-            .setNativeReplicationEnabled(true)
-            .setActiveActiveReplicationEnabled(true)
-            .setNativeReplicationSourceFabric(dcNames[0])).close();
-
-    try (VenicePushJob batchPush = new VenicePushJob("repush-large-batch-v1", batchProps)) {
-      batchPush.run();
-    }
-    waitForVersion(storeName, 1);
-
-    // Repush with chunking enabled + direct control messages
-    Properties repushProps = buildRepushProps(storeName, inputDirPath);
-    repushProps.setProperty(SEND_CONTROL_MESSAGES_DIRECTLY, String.valueOf(sendDirectControlMessage));
-    try (VenicePushJob repushJob = new VenicePushJob("repush-large-repush", repushProps)) {
-      repushJob.run();
-      repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-        StageSummary chunkAssembly = snapshot.getStage("chunk_assembly");
-        assertNotNull(chunkAssembly, "chunk_assembly should be registered for chunked repush");
-        assertTrue(
-            chunkAssembly.getRecordsIn() > numberOfRecords,
-            "chunk_assembly recordsIn (" + chunkAssembly.getRecordsIn() + ") should exceed " + numberOfRecords);
-        assertEquals(chunkAssembly.getRecordsOut(), numberOfRecords, "chunk_assembly recordsOut");
-        assertTrue(chunkAssembly.getBytesIn() > 0, "chunk_assembly bytesIn should be > 0");
-        assertTrue(chunkAssembly.getBytesOut() > 0, "chunk_assembly bytesOut should be > 0");
-        assertSinkStage(snapshot, "kafka_write", numberOfRecords);
-        assertNull(snapshot.getStage("compaction"), "compaction should not be registered when chunking is enabled");
-        assertNull(snapshot.getStage("compression_reencode"));
-        assertNull(snapshot.getStage("ttl_filter"));
-        LOGGER.info("Chunked repush metrics:\n{}", snapshot.getFormattedReport());
-      });
-    }
-    waitForVersion(storeName, 2);
-
-    // Verify large values are present and correct
-    VeniceClusterWrapper cluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-    try (AvroGenericStoreClient<String, Object> client = ClientFactory.getAndStartGenericAvroClient(
-        ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(cluster.getRandomRouterURL()))) {
-      TestUtils.waitForNonDeterministicAssertion(ASSERTION_TIMEOUT_MS, TimeUnit.MILLISECONDS, () -> {
-        for (int i = 0; i < numberOfRecords; i++) {
-          int expectedSize = largeValueSize / numberOfRecords * (i + 1);
-          Object value = client.get(Integer.toString(i)).get();
-          assertNotNull(value, "Key " + i + " is null after chunked repush");
-          assertEquals(value.toString().length(), expectedSize, "Value size mismatch for key " + i);
-        }
-      });
-    }
-  }
-
-  /**
-   * Test KIF repush with both value chunking AND RMD chunking.
-   * Streams partial updates to 3 keys, each accumulating ~2MB of float array data (value > 950KB threshold)
-   * and ~4MB of per-element RMD timestamps (RMD > 950KB threshold). The repush chunk assembler must
-   * reassemble both value chunks and RMD chunks correctly.
-   * Exercises: chunk_assembly (value + RMD) + kafka_write stages.
-   */
-  @Test(timeOut = TEST_TIMEOUT_MS * 3)
-  public void testRepushWithValueAndRmdChunking() {
-    String storeName = Utils.getUniqueString("repush-rmd-chunking");
-    String parentControllerUrl = multiRegionMultiClusterWrapper.getControllerConnectString();
-    Schema valueSchema = AvroCompatibilityHelper.parse(loadFileAsString("CollectionRecordV1.avsc"));
-    Schema partialUpdateSchema = WriteComputeSchemaConverter.getInstance().convertFromValueRecordSchema(valueSchema);
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      assertCommand(
-          parentControllerClient
-              .createNewStore(storeName, "test_owner", STRING_SCHEMA.toString(), valueSchema.toString()));
-      UpdateStoreQueryParams storeParams =
-          new UpdateStoreQueryParams().setStorageQuotaInByte(Store.UNLIMITED_STORAGE_QUOTA)
-              .setPartitionCount(1)
-              .setWriteComputationEnabled(true)
-              .setActiveActiveReplicationEnabled(true)
-              .setChunkingEnabled(true)
-              .setRmdChunkingEnabled(true)
-              .setNativeReplicationEnabled(true)
-              .setNativeReplicationSourceFabric(dcNames[0])
-              .setHybridRewindSeconds(10L)
-              .setHybridOffsetLagThreshold(2L);
-      assertFalse(parentControllerClient.retryableRequest(5, c -> c.updateStore(storeName, storeParams)).isError());
-
-      // Empty push v1 to bootstrap the store
-      VersionCreationResponse response = parentControllerClient.emptyPush(storeName, "test_push_id", 1000);
-      assertEquals(response.getVersion(), 1);
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-
-      // Stream partial updates to 3 keys. Each key gets 50 updates x 10,000 floats = 500K floats.
-      // Value: 500K floats x 4 bytes = ~2MB (exceeds 950KB chunk threshold)
-      // RMD: 500K per-element timestamps x 8 bytes = ~4MB (exceeds 950KB chunk threshold)
-      int numKeys = 3;
-      int updatesPerKey = 50;
-      int floatsPerUpdate = 10_000;
-      VeniceClusterWrapper veniceCluster = childDatacenters.get(0).getClusters().get(CLUSTER_NAME);
-
-      try (VeniceSystemProducer veniceProducer = getSamzaProducer(veniceCluster, storeName, Version.PushType.STREAM)) {
-        for (int k = 0; k < numKeys; k++) {
-          String key = "key_" + k;
-          for (int u = 0; u < updatesPerKey; u++) {
-            UpdateBuilderImpl updateBuilder = new UpdateBuilderImpl(partialUpdateSchema);
-            updateBuilder.setNewFieldValue("name", "record_" + k);
-            java.util.List<Float> newEntries = new java.util.ArrayList<>(floatsPerUpdate);
-            for (int j = 0; j < floatsPerUpdate; j++) {
-              newEntries.add((float) (u * floatsPerUpdate + j));
-            }
-            updateBuilder.setElementsToAddToListField("floatArray", newEntries);
-            GenericRecord partialUpdateRecord = updateBuilder.build();
-            sendStreamingRecord(veniceProducer, storeName, key, partialUpdateRecord, (long) (u + 1));
-          }
-        }
-      }
-
-      // Verify data is ingested before repush (120s timeout: 150 partial updates take time to ingest)
-      try (AvroGenericStoreClient<String, Object> reader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(120, TimeUnit.SECONDS, true, () -> {
-          for (int k = 0; k < numKeys; k++) {
-            Object value = reader.get("key_" + k).get();
-            assertNotNull(value, "key_" + k + " should exist after streaming updates");
-          }
-        });
-      }
-
-      // KIF repush via Spark — exercises chunk_assembly for both value and RMD chunks
-      Properties repushProps = buildRepushProps(storeName, "dummyInputPath");
-      repushProps.put(REWIND_TIME_IN_SECONDS_OVERRIDE, 0);
-      try (VenicePushJob repushJob = new VenicePushJob("repush-rmd-chunking", repushProps)) {
-        repushJob.run();
-        repushJob.getStageMetricsSnapshot().ifPresent(snapshot -> {
-          StageSummary chunkAssembly = snapshot.getStage("chunk_assembly");
-          assertNotNull(chunkAssembly, "chunk_assembly should be registered (both value and RMD are chunked)");
-          // Input records include value chunks + RMD chunks + manifests, far exceeding numKeys
-          assertTrue(
-              chunkAssembly.getRecordsIn() > numKeys,
-              "chunk_assembly recordsIn (" + chunkAssembly.getRecordsIn() + ") should exceed " + numKeys);
-          assertEquals(chunkAssembly.getRecordsOut(), numKeys, "chunk_assembly recordsOut should equal " + numKeys);
-          assertTrue(chunkAssembly.getBytesIn() > 0, "chunk_assembly bytesIn should be > 0");
-          assertTrue(chunkAssembly.getBytesOut() > 0, "chunk_assembly bytesOut should be > 0");
-          assertSinkStage(snapshot, "kafka_write", numKeys);
-          assertNull(snapshot.getStage("compaction"));
-          LOGGER.info("RMD chunking repush metrics:\n{}", snapshot.getFormattedReport());
-        });
-      }
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 2),
-          parentControllerClient,
-          60,
-          TimeUnit.SECONDS);
-
-      // Verify data survives the repush
-      try (AvroGenericStoreClient<String, Object> reader = ClientFactory.getAndStartGenericAvroClient(
-          ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(veniceCluster.getRandomRouterURL()))) {
-        TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, true, () -> {
-          for (int k = 0; k < numKeys; k++) {
-            GenericRecord value = (GenericRecord) reader.get("key_" + k).get();
-            assertNotNull(value, "key_" + k + " should survive repush");
-            assertEquals(value.get("name").toString(), "record_" + k);
-          }
-        });
-      }
-    }
-  }
-
-  // ==================================================================================
-  // Helpers for migrated tests
-  // ==================================================================================
-
-  /**
-   * Build standard repush properties targeting dc-0 with Spark engine.
-   */
-  private Properties buildRepushProps(String storeName, String inputDirPath) {
-    Properties props =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    props.put(SEND_CONTROL_MESSAGES_DIRECTLY, true);
-    props.put(SOURCE_KAFKA, "true");
-    props.put(KAFKA_INPUT_FABRIC, dcNames[0]);
-    props.put(KAFKA_INPUT_MAX_RECORDS_PER_MAPPER, "5000");
-    props.setProperty(DATA_WRITER_COMPUTE_JOB_CLASS, DataWriterSparkJob.class.getCanonicalName());
-    return props;
-  }
-
-  /**
-   * Wait for a version to become current in all DCs.
-   */
-  private void waitForVersion(String storeName, int expectedVersion) {
-    String parentControllerUrl = multiRegionMultiClusterWrapper.getControllerConnectString();
-    try (ControllerClient controllerClient = new ControllerClient(CLUSTER_NAME, parentControllerUrl)) {
-      TestUtils.waitForNonDeterministicAssertion(60, TimeUnit.SECONDS, true, true, () -> {
-        for (int version: controllerClient.getStore(storeName).getStore().getColoToCurrentVersions().values()) {
-          assertEquals(version, expectedVersion, "All DCs should be on v" + expectedVersion);
-        }
-      });
-    }
-  }
-
-  /**
-   * Assert that a stage exists with exact record counts and positive byte/time values.
-   */
-  private void assertStage(
-      StageMetricsRegistry.Snapshot snapshot,
-      String stageName,
-      long expectedRecordsIn,
-      long expectedRecordsOut) {
-    StageSummary stage = snapshot.getStage(stageName);
-    assertNotNull(stage, stageName + " stage should be registered");
-    assertEquals(stage.getRecordsIn(), expectedRecordsIn, stageName + " recordsIn");
-    assertEquals(stage.getRecordsOut(), expectedRecordsOut, stageName + " recordsOut");
-    assertTrue(stage.getBytesIn() > 0, stageName + " bytesIn should be > 0");
-    assertTrue(stage.getBytesOut() > 0, stageName + " bytesOut should be > 0");
-    assertTrue(stage.getTimeNs() > 0, stageName + " timeNs should be > 0");
-  }
-
-  /**
-   * Assert that a sink stage (kafka_write) exists with exact input record count.
-   * Sink stages have recordsOut=0 and bytesOut=0 because the partition writer consumes all rows.
-   */
-  private void assertSinkStage(StageMetricsRegistry.Snapshot snapshot, String stageName, long minExpectedRecordsIn) {
-    StageSummary stage = snapshot.getStage(stageName);
-    assertNotNull(stage, stageName + " stage should be registered");
-    // kafka_write counts data records + spray-all-partitions synthetic records, so use >=
-    assertTrue(
-        stage.getRecordsIn() >= minExpectedRecordsIn,
-        stageName + " recordsIn (" + stage.getRecordsIn() + ") should be >= " + minExpectedRecordsIn);
-    assertTrue(stage.getBytesIn() > 0, stageName + " bytesIn should be > 0");
-    assertTrue(stage.getTimeNs() > 0, stageName + " timeNs should be > 0");
-  }
-
-  /**
-   * Verify that batch data records [1..recordCount] are present and correct in the specified DC.
-   * Keys are string representations of integers, values are "test_name_{key}".
-   */
-  private void verifyBatchData(String storeName, int recordCount, int dcIndex) {
-    VeniceClusterWrapper cluster = childDatacenters.get(dcIndex).getClusters().get(CLUSTER_NAME);
-    String routerUrl = cluster.getRandomRouterURL();
-    TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-      try (AvroGenericStoreClient<String, Object> client = ClientFactory
-          .getAndStartGenericAvroClient(ClientConfig.defaultGenericClientConfig(storeName).setVeniceURL(routerUrl))) {
-        for (int i = 1; i <= recordCount; i++) {
-          Object value = client.get(Integer.toString(i)).get();
-          assertNotNull(value, "Key " + i + " is null in dc-" + dcIndex);
-          assertEquals(value.toString(), "test_name_" + i);
-        }
-      }
-    });
-  }
-
-  private GenericRecord readValue(AvroGenericStoreClient<Object, Object> storeReader, String key)
-      throws ExecutionException, InterruptedException {
-    return (GenericRecord) storeReader.get(key).get();
-  }
-
-  private void validateRmdData(
-      RmdSerDe rmdSerDe,
-      String kafkaTopic,
-      String key,
-      Consumer<RmdWithValueSchemaId> rmdDataValidationFlow) {
-    for (VeniceServerWrapper serverWrapper: multiRegionMultiClusterWrapper.getChildRegions()
-        .get(0)
-        .getClusters()
-        .get("venice-cluster0")
-        .getVeniceServers()) {
-      StorageEngine storageEngine = serverWrapper.getVeniceServer().getStorageService().getStorageEngine(kafkaTopic);
-      assertNotNull(storageEngine);
-      // RMD may not be immediately available after repush (e.g., with ZSTD_WITH_DICT compression where
-      // dictionary-based decompression/recompression can delay RMD persistence). Retry until available.
-      TestUtils.waitForNonDeterministicAssertion(10, TimeUnit.SECONDS, true, true, () -> {
-        ValueRecord result = SingleGetChunkingAdapter
-            .getReplicationMetadata(storageEngine, 0, serializeStringKeyToByteArray(key), true, null);
-        // Avoid assertion failure logging massive RMD record.
-        assertNotNull(result, "RMD should exist for key: " + key);
-        byte[] rmdBytes = result.serialize();
-        RmdWithValueSchemaId rmdWithValueSchemaId = new RmdWithValueSchemaId();
-        rmdSerDe.deserializeValueSchemaIdPrependedRmdBytes(rmdBytes, rmdWithValueSchemaId);
-        rmdDataValidationFlow.accept(rmdWithValueSchemaId);
-      });
-    }
-  }
-
-  private byte[] serializeStringKeyToByteArray(String key) {
-    Utf8 utf8Key = new Utf8(key);
-    DatumWriter<Utf8> writer = new GenericDatumWriter<>(Schema.create(Schema.Type.STRING));
-    ByteArrayOutputStream out = new ByteArrayOutputStream();
-    BinaryEncoder encoder = AvroCompatibilityHelper.newBinaryEncoder(out);
-    try {
-      writer.write(utf8Key, encoder);
-      encoder.flush();
-    } catch (IOException e) {
-      throw new RuntimeException("Failed to write input: " + utf8Key + " to binary encoder", e);
-    }
-    return out.toByteArray();
-  }
-}
