+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/endToEnd/TestRepushTTL.java
@@ -1,0 +1,11 @@
+package com.linkedin.venice.endToEnd;
+
+/**
+ * TTL repush integration tests using the MR (MapReduce) engine.
+ */
+public class TestRepushTTL extends AbstractTestRepushTTL {
+  @Override
+  protected boolean useSpark() {
+    return false;
+  }
+}
