+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/endToEnd/TestRepushTTLSpark.java
@@ -1,0 +1,11 @@
+package com.linkedin.venice.endToEnd;
+
+/**
+ * TTL repush integration tests using the Spark engine.
+ */
+public class TestRepushTTLSpark extends AbstractTestRepushTTL {
+  @Override
+  protected boolean useSpark() {
+    return true;
+  }
+}
